removed post request from the server.js file, and created a routes folder w the stripe post request endpoint
